refine code structure and style
